feat(book): extract CommentBox + AutoPasteFailedDialog into injected-ui
CommentBox only needed a small refactor: it read
activatedSiteAdaptorUI!.networkIdentifier directly for the Minds-specific
width variant and used useLingui() for its placeholder - both are now
site/placeholder props, computed by the container
(site-adaptor-infra/defaults/inject/CommentBox.tsx).

AutoPasteFailedDialog was already close to pure; the only import at risk
of pulling @masknet/shared's barrel into injected-ui was its use of
Image for the pasted screenshot preview. Since that image is always a
local blob URL (never a network fetch), Image's retry/fallback behavior
doesn't apply here, so it's replaced with a plain <img>. The download
button's saveFileFromUrl call and the DraggableDiv positioning wrapper
stay in the container (packages/mask), which now passes onDownload as
a prop.

Adds a date-fns dependency to packages/injected-ui (used for the
suggested download file name) and its pnpm-lock.yaml entry.

Adds CommentBox demos for Twitter/Minds (showing the width variants)
and an AutoPasteFailedDialog demo under apps/book/src/demos/injection/.

Follow-up to #12445, #12446, #12447, #12449, and the not-yet-merged
WizardDialog/PinExtension batch.
